requirements.txt
dash
plotly
pandas
gunicorn
azure-identity
azure-storage-blob
python-dotenv
azure-storage-file-datalake